refact(exporter): extract EC2 logic to a dedicated file
