+++ internal/app/exporter/helper.go
@@ -2,33 +2,7 @@ package exporter
 
 import (
 	"regexp"
-
-	"github.com/qonto/prometheus-rds-exporter/internal/app/rds"
-	"golang.org/x/exp/slices"
 )
-
-func getUniqTypeAndIdentifiers(instances map[string]rds.RdsInstanceMetrics) ([]string, []string) {
-	var (
-		instanceTypes       []string
-		instanceIdentifiers []string
-	)
-
-	for dbinstanceName := range instances {
-		instanceClass := instances[dbinstanceName].DBInstanceClass
-		if !slices.Contains(instanceTypes, instanceClass) {
-			instanceTypes = append(instanceTypes, instanceClass)
-		}
-
-		if !slices.Contains(instanceIdentifiers, dbinstanceName) {
-			instanceIdentifiers = append(instanceIdentifiers, dbinstanceName)
-		}
-	}
-
-	// Remove incompatible instance types
-	instanceTypes = RemoveElementsByValue(instanceTypes, []string{"db.serverless"})
-
-	return instanceIdentifiers, instanceTypes
-}
 
 func ClearPrometheusLabel(str string) string {
 	// Prometheus metric names may contain ASCII letters, digits, underscores, and colons.
